requirements.txt
Django==2.1.7
djangorestframework==3.9.2
pandas==0.24.2
structlog==19.1.0
